Modify-1 : 하나의 파일에서는 하나의 컴포넌트만 관리하도록 구현
