+++ public/data/feed.json
@@ -8,8 +8,7 @@
     "likeNum": "10명",
     "feedContent": "우 나 핫걸 좋은 거 보고 갈래?💦",
     "commentId": "_sentiers",
-    "fixedComment": "라인이 예술이네요",
-    "newComment": []
+    "fixedComment": "라인이 예술이네요"
   },
   {
     "id": 2,
@@ -20,8 +19,7 @@
     "likeNum": "2500명",
     "feedContent": "도촬 전문가 무섭내;;",
     "commentId": "welshicorgi_messi",
-    "fixedComment": "호우 섹시하다",
-    "newComment": []
+    "fixedComment": "호우 섹시하다"
   },
   {
     "id": 3,
@@ -32,7 +30,6 @@
     "likeNum": "25000명",
     "feedContent": "Hard day",
     "commentId": "hot_girl",
-    "fixedComment": "취향 진짜 까다로워 보이는 남미 부자 같아요",
-    "newComment": []
+    "fixedComment": "취향 진짜 까다로워 보이는 남미 부자 같아요"
   }
 ]
